% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Numeric count matrix of example data.}
\format{A numeric matrix with five genes observed across 70 experiments. The first 35 are tumor samples,
  the second 35 are normal samples.}
\source{
Akin, O. and Elnajjar, P. and Heller, M. and Jarosz, R. and Erickson, B. J. and Kirk,
  S. and ... Filippini, J. (2016) Data from The Cancer Genome Atlas Kidney Renal Clear Cell Carcinoma
  (TCGA-KIRC) collection. The Cancer Imaging Archive.
}
\usage{
example_data
}
\description{
The first several rows of a dataset from the Cancer Genome Atlas Kidney Renal Clear Cell Carcinoma
data collection, consisting of 5 genes from a dataset of raw RNA-Seq counts from 35 tumor and 35 normal samples.
}
\keyword{datasets}
